拆解JS檔
