+++ pages/checkout.html
@@ -578,6 +578,8 @@
     <!-- GSAP -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/ScrollTrigger.min.js"></script>
-    <script type="module" src="../main.js"></script>
+    <script type="module" src="../js/checkout.js"></script>
+    <script type="module" src="../js/stickyFooterScrollTrigger.js"></script>
+    <script type="module" src="../js/main.js"></script>
   </body>
 </html>
