+++ pages/order-complete.html
@@ -162,6 +162,6 @@
     <!-- 腳本引入 -->
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/aos@2.3.4/dist/aos.js"></script>
-    <script type="module" src="../main.js"></script>
+    <script type="module" src="../js/main.js"></script>
   </body>
 </html>
