+++ pages/product-detail.html
@@ -407,6 +407,7 @@
     <%- include('./layout/footer'); -%>
     <!-- 腳本引入 -->
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-    <script type="module" src="../main.js"></script>
+    <script type="module" src="../js/product-detail.js"></script>
+    <script type="module" src="../js/main.js"></script>
   </body>
 </html>
