+++ pages/product-list.html
@@ -491,7 +491,7 @@
      <!-- 腳本引入 -->
  
      <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-     <script type="module" src="../main.js"></script>
+     <script type="module" src="../js/main.js"></script>
      <script src="https://cdn.jsdelivr.net/npm/aos@2.3.4/dist/aos.js"></script>
      <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.5/gsap.min.js"></script>
 </body>
